Add real data (#3)
This will need a more realistic dataset. 

As a first pass, I've taken the group email and used everyone there as a
feeder for the `names.csv` file. I've also used realistic dates as to
when people joined.

I've voided a bunch of people that I don't expect to be used.